// src/protos/top_pb.ts
// @generated by protoc-gen-es v2.12.1 with parameter "target=ts"
// @generated from file top.proto (syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { ScrollerData } from "./scroller_pb";
import { file_scroller } from "./scroller_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file top.proto.
 */
export const file_top: GenFile = /*@__PURE__*/
  fileDesc("Cgl0b3AucHJvdG8iXgoVVG9wU2Nyb2xsZXJCdXR0b25EYXRhEhsKBGRhdGEYASABKAsyDS5TY3JvbGxlckRhdGESDAoEdHlwZRgCIAEoDRIRCgRwYWdlGAMgASgNSACIAQFCBwoFX3BhZ2UiNQoWVG9wc1Njcm9sbGVyQnV0dG9uRGF0YRIbCgRkYXRhGAEgASgLMg0uU2Nyb2xsZXJEYXRhYgZwcm90bzM", [file_scroller]);

/**
 * @generated from message TopScrollerButtonData
 */
export type TopScrollerButtonData = Message<"TopScrollerButtonData"> & {
  /**
   * @generated from field: ScrollerData data = 1;
   */
  data?: ScrollerData | undefined;

  /**
   * @generated from field: uint32 type = 2;
   */
  type: number;

  /**
   * @generated from field: optional uint32 page = 3;
   */
  page?: number | undefined;
};

/**
 * Describes the message TopScrollerButtonData.
 * Use `create(TopScrollerButtonDataSchema)` to create a new message.
 */
export const TopScrollerButtonDataSchema: GenMessage<TopScrollerButtonData> = /*@__PURE__*/
  messageDesc(file_top, 0);

/**
 * @generated from message TopsScrollerButtonData
 */
export type TopsScrollerButtonData = Message<"TopsScrollerButtonData"> & {
  /**
   * @generated from field: ScrollerData data = 1;
   */
  data?: ScrollerData | undefined;
};

/**
 * Describes the message TopsScrollerButtonData.
 * Use `create(TopsScrollerButtonDataSchema)` to create a new message.
 */
export const TopsScrollerButtonDataSchema: GenMessage<TopsScrollerButtonData> = /*@__PURE__*/
  messageDesc(file_top, 1);